pyproject.toml
[project]
name = "nextgen-uia-private"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.1",
    "filelock==3.20.0",
    "fsspec==2025.9.0",
    "ftfy>=6.3.1",
    "jinja2==3.1.6",
    "markupsafe==3.0.3",
    "matplotlib>=3.10.7",
    "monai>=1.5.1",
    "mpmath==1.3.0",
    "networkx==3.5",
    "numpy==2.3.4",
    "nvidia-cublas-cu12==12.8.4.1",
    "nvidia-cuda-cupti-cu12==12.8.90",
    "nvidia-cuda-nvrtc-cu12==12.8.93",
    "nvidia-cuda-runtime-cu12==12.8.90",
    "nvidia-cudnn-cu12==9.10.2.21",
    "nvidia-cufft-cu12==11.3.3.83",
    "nvidia-cufile-cu12==1.13.1.3",
    "nvidia-curand-cu12==10.3.9.90",
    "nvidia-cusolver-cu12==11.7.3.90",
    "nvidia-cusparse-cu12==12.5.8.93",
    "nvidia-cusparselt-cu12==0.7.1",
    "nvidia-nccl-cu12==2.27.5",
    "nvidia-nvjitlink-cu12==12.8.93",
    "nvidia-nvshmem-cu12==3.3.20",
    "nvidia-nvtx-cu12==12.8.90",
    "nvitop>=1.6.0",
    "open-clip-torch>=3.2.0",
    "optuna>=4.5.0",
    "pandas>=2.3.3",
    "pillow==12.0.0",
    "regex>=2025.10.23",
    "ruff>=0.14.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "setuptools==80.9.0",
    "sympy==1.14.0",
    "tensorboard>=2.20.0",
    "torch==2.9.0",
    "torchaudio==2.9.0",
    "torchmetrics>=1.8.2",
    "torchvision==0.24.0",
    "tqdm==4.67.1",
    "transformers>=4.57.1",
    "triton==3.5.0",
    "typing-extensions==4.15.0",
]
